+++ src/components/Navbar/molecules/MainNav.tsx
@@ -7,8 +7,7 @@ import { cn } from "~/utils/cn";
 import { Icons } from "~/components/Icons";
 import { MobileNav } from "./MobileNav";
 import { type Route } from "next";
-import Icon from "~/assets/icon.png";
-import Image from "next/image";
+import { FeastQRLogo } from "./FeastQRLogo";
 
 export type NavItem = {
   title: string | React.ReactNode;
@@ -29,10 +28,7 @@ export function MainNav({ items, children }: MainNavProps) {
   return (
     <div className="flex gap-6 md:gap-10">
       <Link href="/" className="hidden items-center space-x-2 md:flex">
-        <div className="flex items-center gap-3">
-          <Image src={Icon} alt="Icon" height={30} />
-          <span className="text-xl font-semibold">FeastQR</span>
-        </div>
+        <FeastQRLogo />
       </Link>
       {items?.length ? (
         <nav className="hidden gap-6 md:flex">
