+++ src/i18n/settings.ts
@@ -1,11 +1,14 @@
-import { type Namespace } from 'i18next';
+import { type Namespace } from "i18next";
 
-export const fallbackLng = 'pl' as const;
-export const languages = [fallbackLng, 'en'] as const;
-export const defaultNamespace: Namespace = 'common';
+export const fallbackLng = "en" as const;
+export const languages = [fallbackLng, "pl"] as const;
+export const defaultNamespace: Namespace = "common";
 export type Language = (typeof languages)[number];
 
-export function getOptions(lng: Language = fallbackLng, ns: Namespace = defaultNamespace) {
+export function getOptions(
+  lng: Language = fallbackLng,
+  ns: Namespace = defaultNamespace,
+) {
   return {
     supportedLngs: languages,
     fallbackLng: languages,
